requirements.txt
loguru==0.7.3
torch_tb_profiler==0.4.3
tabulate==0.9.0
matplotlib==3.10.1
transformers==4.51.1
datasets==3.4.1
peft==0.15.0
accelerate==1.5.2
flash-attn==2.7.4.post1
huggingface-hub==0.30.2
evaluate==0.4.3
einops==0.8.1
wandb==0.19.8
tensorboard==2.19.0
pandas==2.2.3
numpy==1.26.4
scipy==1.15.2
hf_transfer==0.1.9
megatron-core==0.11.0
liger-kernel==0.5.6
seaborn==0.13.2
mypy_extensions==1.0.0
trl==0.9.6
pre-commit==4.1.0
huggingface-hub==0.30.2